+++ .tools/test-bafu-hydro-ch.php
@@ -113,6 +113,9 @@ class IPSModule
     public function SendDebug($s, $m, $f)
     {
     }
+    public function LogMessage($Message, $Type)
+    {
+    }
     public function UpdateFormField($n, $k, $v)
     {
     }
@@ -129,6 +132,13 @@ class IPSModule
 
 const VARIABLETYPE_STRING = 3;
 const VARIABLETYPE_INTEGER = 1;
+const KL_MESSAGE = 10201;
+const KL_SUCCESS = 10202;
+const KL_NOTIFY = 10203;
+const KL_WARNING = 10204;
+const KL_ERROR = 10205;
+const KL_DEBUG = 10206;
+const KL_CUSTOM = 10207;
 function IPS_VariableProfileExists(string $name): bool
 {
     return true;